Don’t track generated `lib/git-hub.d/completion`
This commit deletes and un-tracks the generated Bash script
`lib/git-hub.d/completion`, which is the output of the Perl script
`tool/generate-completion.pl` when the latter is fed `doc/git-hub.swim`
via its standard input stream.

Tracking generated files in a Git repository is generally bad practice —
<https://sethrobertson.github.io/GitBestPractices/#donot> — and I would
not consider the file in question to be difficult to generate.

This commit also updates the documentation document `doc/git-hub.swim`
to instruct users, rather than to `source` (in Bash)
`lib/git-hub.d/completion`, to `eval` (in Bash) the output of
`./tool/generate-completion.pl < doc/git-hub.swim`.

This commit does place additional burden on the end-user, but I would
consider this burden to be fairly minor, to the tune of
copying-and-pasting an additional 33 characters, and I intend to
implement additional tooling to make the whole process easier than
before.